Refactoring around how store context is handled
The store code being pulled out of a cookie was already happening in the
app init method and not needed. Any time the store code was in the cookie
it was basically a noop. This code was removed and now instead the startFrontendSession
method will toss an exception (403 HTTP code) if you are in an admin store context.

This change has the benefit of respecting what Magento has already set the store to without needing special checks.

The session login module has been updated to handle this exception in an intelligent manner
by not handling it and just re-throwing it so the API handler get return a proper 403 response.