Refactor the adapter to init frontend sessions
The previous behaviour was to only init the session if the correct cookie was found.
This worked fine for logged in customers, but guest customers had problems.
If you used the an API as a guest customer and NO SESSION, then a session would be created
without the proper session name. This resulted in a lost session once you visited the site.

The new behaviour will always initialize a frontend session for non-admin store contexts.
This results in a proper session name for guest users.