Admin check is unnecessary
True admin URLs always begin with the adminhtml front name and so
something beginning with `/api/rest` is definitely not the admin store.
Conceptually the API has no store but Mage needs a store context.
`api.php` always initialises the admin store.
So there is no point in checking for admin.
Also, a logged in admin does not negate a customer session.